requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
channels==4.0.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==4.1.7
django-redis==5.2.0
django-rest==0.8.7
djangorestframework==3.14.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
install==1.3.5
kombu==5.2.4
multidict==6.0.4
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.4
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
yarl==1.8.2
